ohto/requirements.txt
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==38.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0
django-bootstrap5==22.1
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-extensions==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
idna==3.4
importlib-metadata==2.0.0
inflection==0.5.1
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.16.0
jupyter_client==7.3.5
jupyter_core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.1.2
jupyterlab_server==2.15.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==0.8.4
mysqlclient==2.0.3
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
notebook==6.4.12
numexpr==2.8.3
numpy==1.23.3
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.20
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
Send2Trash==1.8.0
setuptools==65.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
stack-data==0.2.0
terminado==0.13.1
testpath==0.6.0
tornado==6.2
traitlets==5.1.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
zipp==3.8.0
